refactor: move type to common
