+++ __tests__/table/golangcilint.test.ts
@@ -1,6 +1,6 @@
 import { GolangCILintTable } from '../../src/table/golangcilint'
 import { GolangCILintSummaryReport } from '../../src/report/type'
-import { Result } from '../../src/type'
+import { Result } from '../../src/common/type'
 import { Align } from '../../src/table/base/type'
 
 describe('GolangCILintTable', () => {
