+++ src/report/golangcilint.ts
@@ -1,4 +1,4 @@
-import { Result, GitHubContext } from '../type'
+import { Result, GitHubContext } from '../common/type'
 import {
   Index,
   GolangCILintSummaryReport,
