+++ src/reporter/factory.ts
@@ -1,5 +1,5 @@
 import { AnyReporter } from './type'
-import { ReporterType, GitHubContext } from '../type'
+import { ReporterType, GitHubContext } from '../common/type'
 import { GolangCILintReporterImpl } from './golangcilint'
 import { GotestsumReporterImpl } from './gotestsum'
 import { JUnitXmlReader } from './reader'
