+++ src/table/result.ts
@@ -1,4 +1,4 @@
-import { Result } from '../type'
+import { Result } from '../common/type'
 
 export function toResult(results: Result[]): Result {
   return results.some(result => result == Result.Failed)
